ai-engine/collectorAI/requirements.txt
flask
flask-cors
requests
Pillow